% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Missing_species.R
\name{Missing_species}
\alias{Missing_species}
\title{Bird Species Observed on eBird but not iNat}
\usage{
Missing_species(user)
}
\arguments{
\item{user}{The target iNaturalist username}
}
\value{
A list of bird species
}
\description{
Find out which species you have observed in eBird but not in iNaturalist
}
\examples{
Species  <- Missing_species("whimbrelbirder")
}
